fixup: refactor module into separate files
Refactors update_function_docs module into separate files to
improve readability.
