+++ scripts/update_function_docs/function_release.go
@@ -11,23 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
-// Usage: update_function_docs <RELEASE_BRANCH>
-//
-// e.g. update_function_docs origin/apply-setters/v0.2
-//
-// The command will checkout the release branch and update the function/example
-// docs with the latest patch version for the release. If the docs are updated
-// then a commit is created with the changes. The manual steps left to the user
-// are to push the commit to a branch and create a pull request.
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"regexp"
 	"strings"
@@ -35,65 +24,6 @@ import (
 	"golang.org/x/mod/semver"
 	"gopkg.in/yaml.v2"
 )
-
-func exitWithErr(err error) {
-	fmt.Fprintf(os.Stderr, "%v\n", err)
-	os.Exit(1)
-}
-
-func runCmd(name string, arg ...string) (string, error) {
-	var stdout bytes.Buffer
-	var stderr bytes.Buffer
-	cmd := exec.Command(name, arg...)
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	fmt.Printf("%s\n", cmd.String())
-	err := cmd.Run()
-	if err != nil {
-		return stdout.String(), fmt.Errorf("%s\n%s", stderr.String(), err)
-	}
-	return stdout.String(), err
-}
-
-func isCleanRepo() bool {
-	_, err := runCmd("git", "diff-index", "--quiet", "HEAD", "--")
-	if err != nil {
-		return false
-	}
-	return true
-}
-
-func gitFetch() error {
-	_, err := runCmd("git", "fetch", "--tags")
-	return err
-}
-
-func gitCheckout(branch string) error {
-	_, err := runCmd("git", "checkout", branch)
-	return err
-}
-
-func gitTag() (string, error) {
-	return runCmd("git", "tag")
-}
-
-func gitAdd() error {
-	_, err := runCmd("git", "add", "-u")
-	return err
-}
-
-func gitCommit(msg string) error {
-	formattedMsg := fmt.Sprintf("\"%s\"", msg)
-	stdout, err := runCmd("git", "commit", "-m", formattedMsg)
-	fmt.Printf("%v\n", stdout)
-	return err
-}
-
-func gitShow() error {
-	stdout, err := runCmd("git", "show")
-	fmt.Printf("%v\n", stdout)
-	return err
-}
 
 var (
 	// pattern of release branches, e.g. apply-setters/v1.0
@@ -263,6 +193,49 @@ func (fr *functionRelease) parseMetadata(examplesPath string) error {
 	return nil
 }
 
+// updateDocs updates all the docs for the functionRelease on the filesystem
+func (fr *functionRelease) updateDocs() error {
+	if err := fr.updateFunctionDoc(); err != nil {
+		return err
+	}
+	if err := fr.updateExampleDocs(); err != nil {
+		return err
+	}
+	return nil
+}
+
+// updateFunctionDoc updates the function docs for the functionRelease
+func (fr *functionRelease) updateFunctionDoc() error {
+	functionReadme := filepath.Join(fr.FunctionPath, "README.md")
+	return fr.updateDoc(functionReadme)
+}
+
+// updateExampleDocs updates the example docs for the functionRelease
+func (fr *functionRelease) updateExampleDocs() error {
+	for _, example := range fr.Examples {
+		exampleReadme := filepath.Join(example.ExamplePath, "README.md")
+		if err := fr.updateDoc(exampleReadme); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// Perform in place search/replace operations on a documentation file
+func (fr *functionRelease) updateDoc(filePath string) error {
+	contents, err := ioutil.ReadFile(filePath)
+	if err != nil {
+		return err
+	}
+	contents = fr.replaceTags(contents)
+	contents = fr.replaceURLs(contents)
+	contents = fr.replaceKptPackages(contents)
+	if err = os.WriteFile(filePath, contents, 0644); err != nil {
+		return err
+	}
+	return nil
+}
+
 // replace tags with patch e.g. apply-setters:v1.0.1, apply-setters/v1.0.1
 func (fr *functionRelease) replaceTags(contents []byte) []byte {
 	tagPattern := regexp.MustCompile(
@@ -296,85 +269,4 @@ func (fr *functionRelease) replaceKptPackages(contents []byte) []byte {
 	contents = kptPkgPattern.ReplaceAll(contents,
 		[]byte(fmt.Sprintf(`${1}${2}@%s/%s${3}`, fr.FunctionName, fr.LatestPatchVersion)))
 	return contents
-}
-
-// Perform in place search/replace operations on a documentation file
-func (fr *functionRelease) updateDoc(filePath string) error {
-	contents, err := ioutil.ReadFile(filePath)
-	if err != nil {
-		return err
-	}
-	contents = fr.replaceTags(contents)
-	contents = fr.replaceURLs(contents)
-	contents = fr.replaceKptPackages(contents)
-	if err = os.WriteFile(filePath, contents, 0644); err != nil {
-		return err
-	}
-	return nil
-}
-
-// updateFunctionDoc updates the function docs for the functionRelease
-func (fr *functionRelease) updateFunctionDoc() error {
-	functionReadme := filepath.Join(fr.FunctionPath, "README.md")
-	return fr.updateDoc(functionReadme)
-}
-
-// updateExampleDocs updates the example docs for the functionRelease
-func (fr *functionRelease) updateExampleDocs() error {
-	for _, example := range fr.Examples {
-		exampleReadme := filepath.Join(example.ExamplePath, "README.md")
-		if err := fr.updateDoc(exampleReadme); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// updateDocs updates all the docs for the functionRelease on the filesystem
-func (fr *functionRelease) updateDocs() error {
-	if err := fr.updateFunctionDoc(); err != nil {
-		return err
-	}
-	if err := fr.updateExampleDocs(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func main() {
-	var err error
-	if len(os.Args) < 2 {
-		exitWithErr(fmt.Errorf("usage: update_function_docs <RELEASE_BRANCH>"))
-	}
-	releaseBranch := os.Args[1]
-	if !isCleanRepo() {
-		exitWithErr(fmt.Errorf("dirty repo"))
-	}
-	if err = gitFetch(); err != nil {
-		exitWithErr(err)
-	}
-	if err = gitCheckout(releaseBranch); err != nil {
-		exitWithErr(err)
-	}
-	fr, err := newFunctionRelease(releaseBranch)
-	if err != nil {
-		exitWithErr(err)
-	}
-	if err = fr.updateDocs(); err != nil {
-		exitWithErr(err)
-	}
-	if isCleanRepo() {
-		exitWithErr(fmt.Errorf("docs up to date"))
-	}
-	if err = gitAdd(); err != nil {
-		exitWithErr(err)
-	}
-	msg := fmt.Sprintf("docs: Update tags for %s/%s/%s",
-		fr.Language, fr.FunctionName, fr.LatestPatchVersion)
-	if err = gitCommit(msg); err != nil {
-		exitWithErr(err)
-	}
-	if err = gitShow(); err != nil {
-		exitWithErr(err)
-	}
 }
